Chore: made Pencil component + renamed to StudentProfile
from StudentProfileRoute -> StudentProfile
